+++ ManagementSystem.Common/Models/Dtos/Accounting/LegerResponseDto.cs
@@ -11,7 +11,7 @@ namespace ManagementSystem.Common.Models.Dtos
         public int? DoccumentNumber { get; set; }
         public string? DoccumentType { get; set; }
         public int? BillId { get; set; }
-        public int? Amount { get; set; }
+        public long? Amount { get; set; }
         public string LegerDescription { get; set; }
         public int? CustomerId { get; set; }
         public string CustomerName { get; set; }
